Moved test into scalar_to_symbol_promotion_test.py
